refactor: extract chat payload boundaries
